Adding "Press any Key" prompt
Adds a "press any key to continue..." prompt at end of script execution. 

This will prevent console window from automatically closing in the case that the user is using "drag and drop" functionality to execute the scripts by dropping a .key or .wallet or .cipher file onto the script as opposed to running from the command line.